Refactor - moved singleton logic to helper function
